created: enderecos repository and him interface's

+++ src/Application/DTOs/CreateProdutoPedidoDTO.cs
@@ -3,5 +3,5 @@ namespace Comaagora_API.src.Application.DTOs;
 public class CreateProdutoPedidoDTO
 {
     public int ProdutoId { get; set; }
-    public int Quantidade { get; set; }
+    public decimal Quantidade { get; set; }
 }
